gemini-multimodal-langchain/requirements.txt
langchain
langchain-core 
langchain-google-vertexai
python-dotenv